refactor: add magic number for constant
