+++ src/view/OutputView.js
@@ -1,6 +1,6 @@
 const ResultContainer = require("../model/ResultContainer");
 const Console = require("./utils/Console");
-const {OUTPUT_MESSAGE} = require("./constants/message")
+const {OUTPUT_MESSAGE, OUTPUT_SIGNAL} = require("./utils/constants/message")
 class OutputView {
   static printResult(finalResult) {
     if (!finalResult instanceof ResultContainer) {
@@ -11,26 +11,26 @@ class OutputView {
   }
 
   static #getFinalResultMessage(finalResult) {
-    return `${OUTPUT_MESSAGE.result}\n${this.#getRoundResultMessage(finalResult.roundResultList)}${this.#getWinnerMessage(finalResult.winnerList)}`
+    return `${OUTPUT_MESSAGE.result}\n${this.#getRoundResultMessage(finalResult.roundResultList)}${this.#getWinnerMessage(finalResult.winnerList)}`;
   }
 
   static #getRoundResultMessage(roundResultList) {
     return roundResultList.reduce((message, roundResult) => {
   
       roundResult.forEach((score, car) => {
-        message += `${car}: ${'-'.repeat(score)}\n`;
+        message += `${car}: ${OUTPUT_SIGNAL.carMovement.repeat(score)}\n`;
       })
 
-      return message + '\n'
+      return message + '\n';
     },'')
   }
 
   static #getWinnerMessage(winnerList) {
-    return winnerList.join(', ') + OUTPUT_MESSAGE.winner
+    return winnerList.join(', ') + OUTPUT_MESSAGE.winner;
   }
 
   static printError(error) {
-    console.log(error.message)
+    console.log(error.message);
   }
 
   static close() {
